refactor: player base
